Add files via upload
This code does not incorporate stationary data along with the time series predictions. After predictions are made, the trend is fit to a polynomial regression model whose parameters can be used to train another machine learning framework that predicts the regression parameters based on features of US counties. Additionally, the prediction framework is incorporated into a single function at the end of the notebook.